Moved CompileRegistry
